requirements.txt
opencv-python>=4.8.0
requests>=2.31.0
numpy>=1.24.0
pyyaml>=6.0
pillow>=10.0.0
urllib3>=2.0.0

# Optional dependencies
# ffmpeg-python>=0.2.0